Add init Ionic example application code commit
This adds the basic structure for the Ionic TTS/STT example application.  
The application has two tabs. The first is the STT example and has a button and text output elements. The second tab is the TTS example and has a button and text display elements. The GUI elements do not have any functionality in this commit.