Add init Ionic example application code commit (#4)
* Add init Ionic example application code commit

This adds the basic structure for the Ionic TTS/STT example application.  
The application has two tabs. The first is the STT example and has a button and text output elements. The second tab is the TTS example and has a button and text display elements. The GUI elements do not have any functionality in this commit.

* Add README setup instructions

This contains the basic setup steps required to get the example application running on Android and IOS devices or emulators.

* Rename project and directory structure

This renames the project to ionic-angular-example and restructures the directories to examples/* for demo applications

* Removed auto-generated comments and rename function

This removes some auto-generated comments and renames captureButtonClicked method to captureSpeechButtonClicked

* Fix e2e test

This fixes the auto-generated e2e test by updating the existing test to look for the proper page name.