Security updates for cordova-paramedic (#29)
* Add Cordova Plugin Testing Framework

This adds the framework for testing the Cognitive Services Cordova Plugin.
To run tests use:
     npm run test:android or npm run testLios

Test(s) can be found in tests/tests.js file

* Add Browser as platform for tests

This adds browser support to the testing framework.
To run browser test:
    npm run test:browser

* Add security updates for cordova-paramedic

This changes the cordova-paramedic package to point at a updated branch with the latest security patches. This branch was created by an automation tool, but has not been excepted into master yet.