Adding work-around steps to ReadMe (#33)
* Add Cordova Plugin Testing Framework

This adds the framework for testing the Cognitive Services Cordova Plugin.
To run tests use:
     npm run test:android or npm run testLios

Test(s) can be found in tests/tests.js file

* Add Browser as platform for tests

This adds browser support to the testing framework.
To run browser test:
    npm run test:browser

* Add steps to Readme for potential error when plugin tests on Windows

This adds work-around steps to the ReadMe file for a potential error when trying to test the plugin for Android devices on the Windows platform. Suggestion is to launch Android device emulator before launching the tests.